Update Specification_bitwise.txt
Make sure developers have general access to literal bitwise ints and floats. Make sure that developers have access to a ternary operator that doesn’t use nearly as many characters as writing out an if-else clause, and which has general availability within statements without being too confusing. (Unless there is an “n” “t” or “=“ or “+” or “-“ or “/“ or “*” adjacent to the closing-bracket-like use of “\”.